A fix for redis3+
node_redis: Deprecated: The HSET command contains a argument of type Cookie.
This is converted to "test1=123; Max-Age=5; Domain=example.com; Path=/" by using .toString() now and will return an error from v.3.0 on.
Please handle this in your code to make sure everything works as you intended it to.